requirements.txt
torch
torchvision
pillow
requests
fastapi
streamlit
uvicorn
python-multipart
matplotlib
seaborn
tqdm